+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 ## Changelog
 
+### Version 5.1.0-alpha6 [WIP]
+
+- Fix iOS bug which would break size of views when video is displayed with controls on a non full-screen React view. [#1931](https://github.com/react-native-community/react-native-video/pull/1931)
+- Fix video dimensions being undefined when playing HLS in ios. [#1992](https://github.com/react-native-community/react-native-video/pull/1992)
+- Add support for audio mix with other apps for iOS. [#1978](https://github.com/react-native-community/react-native-video/pull/1978)
+- Properly implement pending seek for iOS. [#1994](https://github.com/react-native-community/react-native-video/pull/1994)
+
+### Version 5.1.0-alpha5
+
+- Add support for react-native Windows Cpp/WinRT [#1893]((https://github.com/react-native-community/react-native-video/pull/1893))
+
+### Version 5.1.0-alpha4
+
+- Fix android play/pause bug related to full-screen mode [#1916](https://github.com/react-native-community/react-native-video/pull/1916)
+
 ### Version 5.1.0-alpha3
 
 - Improve Android Audio Focus [#1897](https://github.com/react-native-community/react-native-video/pull/1897)